arrumando as classe dao, as classes de modelo, testando e arrumando as classes de controle de requisiçoes e colocando algumas validaçoes e a classe de controle de cursos testando e arrumando algumas coisas no modelo que fazia ela não cadastrar

+++ src/main/java/br/edu/ifrs/restinga/requisicoes/controle/AlunosControle.java
@@ -57,7 +57,7 @@ public class AlunosControle {
 
     @RequestMapping(path = "/{id}", method = RequestMethod.DELETE)
     @ResponseStatus(HttpStatus.OK)
-    public void apagar(@PathVariable int id) {
+    public void apagar(@PathVariable Long id) {
         if (alunoDAO.existsById(id)) {
             alunoDAO.deleteById(id);
         } else {
