+++ src/main/java/br/edu/ifrs/restinga/requisicoes/dao/UsuarioDAO.java
@@ -10,6 +10,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UsuarioDAO extends JpaRepository<Usuario, Integer> {
+public interface UsuarioDAO extends JpaRepository<Usuario, Long> {
     
 }
